Fix check group handling, refactor scenario framework
